GC_HACKATHON/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
requests
pathway
subprocess
python-dotenv
phi 